+++ ft_printf/ft_puthexa_lower.c
@@ -6,13 +6,13 @@
 /*   By: welow <welow@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/27 12:08:20 by welow             #+#    #+#             */
-/*   Updated: 2023/10/27 14:15:11 by welow            ###   ########.fr       */
+/*   Updated: 2023/11/01 21:56:32 by welow            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_puthexa_lower(unsigned int nb)
+int	ft_puthexa_lower(unsigned long long nb)
 {
 	int	count;
